+++ index.js
@@ -21,7 +21,7 @@ import Stack from "./src/components/stack";
 import Video from "./src/components/video";
 import formatCredits from "./src/utils/format-credits";
 import formatPowaVideoEmbed from "./src/utils/format-powa-video-embed";
-import getImageFromANS from "./src/utils/get-image-from-ans/getImageFromANS";
+import getImageFromANS from "./src/utils/get-image-from-ans";
 import isServerSide from "./src/utils/is-server-side/isServerSide";
 import useInterval from "./src/utils/use-interval/useInterval";
